add adc/iwdt module and fix gpio ana func bug
Test Plan: No

Reviewers: mingzhou, xtzhang

Reviewed By: mingzhou

Differential Revision: http://10.10.20.8/D196

+++ dev/peripheral/crypt/reg_lscrypt_type.h
@@ -32,121 +32,121 @@ typedef struct
 
 enum CRYPT_REG_DATA0_FIELD
 {
-    CRYPT_DATA0_MASK = 0xffffffff,
+    CRYPT_DATA0_MASK = (int)0xffffffff,
     CRYPT_DATA0_POS = 0,
 };
 
 enum CRYPT_REG_DATA1_FIELD
 {
-    CRYPT_DATA1_MASK = 0xffffffff,
+    CRYPT_DATA1_MASK = (int)0xffffffff,
     CRYPT_DATA1_POS = 0,
 };
 
 enum CRYPT_REG_DATA2_FIELD
 {
-    CRYPT_DATA2_MASK = 0xffffffff,
+    CRYPT_DATA2_MASK = (int)0xffffffff,
     CRYPT_DATA2_POS = 0,
 };
 
 enum CRYPT_REG_DATA3_FIELD
 {
-    CRYPT_DATA3_MASK = 0xffffffff,
+    CRYPT_DATA3_MASK = (int)0xffffffff,
     CRYPT_DATA3_POS = 0,
 };
 
 enum CRYPT_REG_KEY0_FIELD
 {
-    CRYPT_KEY0_MASK = 0xffffffff,
+    CRYPT_KEY0_MASK = (int)0xffffffff,
     CRYPT_KEY0_POS = 0,
 };
 
 enum CRYPT_REG_KEY1_FIELD
 {
-    CRYPT_KEY1_MASK = 0xffffffff,
+    CRYPT_KEY1_MASK = (int)0xffffffff,
     CRYPT_KEY1_POS = 0,
 };
 
 enum CRYPT_REG_KEY2_FIELD
 {
-    CRYPT_KEY2_MASK = 0xffffffff,
+    CRYPT_KEY2_MASK = (int)0xffffffff,
     CRYPT_KEY2_POS = 0,
 };
 
 enum CRYPT_REG_KEY3_FIELD
 {
-    CRYPT_KEY3_MASK = 0xffffffff,
+    CRYPT_KEY3_MASK = (int)0xffffffff,
     CRYPT_KEY3_POS = 0,
 };
 
 enum CRYPT_REG_KEY4_FIELD
 {
-    CRYPT_KEY4_MASK = 0xffffffff,
+    CRYPT_KEY4_MASK = (int)0xffffffff,
     CRYPT_KEY4_POS = 0,
 };
 
 enum CRYPT_REG_KEY5_FIELD
 {
-    CRYPT_KEY5_MASK = 0xffffffff,
+    CRYPT_KEY5_MASK = (int)0xffffffff,
     CRYPT_KEY5_POS = 0,
 };
 
 enum CRYPT_REG_KEY6_FIELD
 {
-    CRYPT_KEY6_MASK = 0xffffffff,
+    CRYPT_KEY6_MASK = (int)0xffffffff,
     CRYPT_KEY6_POS = 0,
 };
 
 enum CRYPT_REG_KEY7_FIELD
 {
-    CRYPT_KEY7_MASK = 0xffffffff,
+    CRYPT_KEY7_MASK = (int)0xffffffff,
     CRYPT_KEY7_POS = 0,
 };
 
 enum CRYPT_REG_IVR0_FIELD
 {
-    CRYPT_IVR0_MASK = 0xffffffff,
+    CRYPT_IVR0_MASK = (int)0xffffffff,
     CRYPT_IVR0_POS = 0,
 };
 
 enum CRYPT_REG_IVR1_FIELD
 {
-    CRYPT_IVR1_MASK = 0xffffffff,
+    CRYPT_IVR1_MASK = (int)0xffffffff,
     CRYPT_IVR1_POS = 0,
 };
 
 enum CRYPT_REG_IVR2_FIELD
 {
-    CRYPT_IVR2_MASK = 0xffffffff,
+    CRYPT_IVR2_MASK = (int)0xffffffff,
     CRYPT_IVR2_POS = 0,
 };
 
 enum CRYPT_REG_IVR3_FIELD
 {
-    CRYPT_IVR3_MASK = 0xffffffff,
+    CRYPT_IVR3_MASK = (int)0xffffffff,
     CRYPT_IVR3_POS = 0,
 };
 
 enum CRYPT_REG_RES0_FIELD
 {
-    CRYPT_RES0_MASK = 0xffffffff,
+    CRYPT_RES0_MASK = (int)0xffffffff,
     CRYPT_RES0_POS = 0,
 };
 
 enum CRYPT_REG_RES1_FIELD
 {
-    CRYPT_RES1_MASK = 0xffffffff,
+    CRYPT_RES1_MASK = (int)0xffffffff,
     CRYPT_RES1_POS = 0,
 };
 
 enum CRYPT_REG_RES2_FIELD
 {
-    CRYPT_RES2_MASK = 0xffffffff,
+    CRYPT_RES2_MASK = (int)0xffffffff,
     CRYPT_RES2_POS = 0,
 };
 
 enum CRYPT_REG_RES3_FIELD
 {
-    CRYPT_RES3_MASK = 0xffffffff,
+    CRYPT_RES3_MASK = (int)0xffffffff,
     CRYPT_RES3_POS = 0,
 };
 
@@ -176,7 +176,7 @@ enum CRYPT_REG_CR_FIELD
     CRYPT_FIFOODR_POS = 13,
     CRYPT_DMAEN_MASK = 0x4000,
     CRYPT_DMAEN_POS = 14,
-    CRYPT_CRYSEL_MASK = 0x80000000,
+    CRYPT_CRYSEL_MASK = (int)0x80000000,
     CRYPT_CRYSEL_POS = 31,
 };
 
@@ -202,7 +202,7 @@ enum CRYPT_REG_ICRF_FIELD
 
 enum CRYPT_REG_FIFO_FIELD
 {
-    CRYPT_FIFO_MASK = 0xffffffff,
+    CRYPT_FIFO_MASK = (int)0xffffffff,
     CRYPT_FIFO_POS = 0,
 };
